+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.7] - 2024-06-14
+
+### Fixed
+
+- datacube cloud masked when using `groupby_date`.
+
+### Changed
+
+- accessors are now more compatible for dataarray and dataset.
+- `ed.clip`accessor now manages first bounding box then clip geometry to be faster.
+- New whittaker function that is ~20 times -faster.
+/!\ Beware, lambda from previous whittaker must be multiplied by 10 000 to have same results.
+
 ## [0.1.6] - 2024-06-05
 
 ### Fixed